Replace std::views::enumerate with index-based loops and change hash type from uint32_t to size_t
Removes C++23 std::views::enumerate usage in ValidateArray.hpp and ValidateCombinators.hpp, replacing with traditional index-based for loops to maintain C++20 compatibility. Changes CachedKey::hash type from uint32_t to size_t in CacheOptimizedStructures.hpp to match platform-native hash function return types.

+++ include/json-query/json-path/internal/CacheOptimizedStructures.hpp
@@ -285,10 +285,10 @@ class KeyInternCache
     struct CachedKey
     {
         QString  key;
-        uint32_t hash;
+        size_t   hash;
         uint32_t accessCount{1};
 
-        CachedKey(const QString& k, uint32_t h) : key(k), hash(h) {}
+        CachedKey(const QString& k, size_t h) : key(k), hash(h) {}
     };
 
     std::vector<CachedKey> cache_;
